requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
einops==0.6.0
filelock==3.8.2
frozenlist==1.3.3
fsspec==2022.11.0
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
lightning-utilities==0.4.2
multidict==6.0.3
numpy==1.24.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.1
python-dateutil==2.8.2
pytorch-lightning==1.8.5.post0
pytorch-metric-learning==1.6.3
pytz==2022.7
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
sacremoses==0.0.53
scikit-learn==1.2.0
scipy==1.9.3
sentencepiece==0.1.97
six==1.16.0
tensorboardX==2.5.1
threadpoolctl==3.1.0
tokenizers==0.10.3
torch==1.13.1
torchmetrics==0.11.0
torchvision==0.14.1
tqdm==4.64.1
transformers==4.14.1
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
--extra-index-url http://download.pytorch.org/whl/cu116
